+++ lib/page/marker/add_images_widget.dart
@@ -5,9 +5,8 @@ import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:insigno_frontend/util/iterable.dart';
 import 'package:insigno_frontend/util/nullable.dart';
+import 'package:insigno_frontend/util/pair.dart';
 import 'package:os_detect/os_detect.dart' as Platform;
-
-import '../util/pair.dart';
 
 class AddImagesWidget extends StatelessWidget {
   static const imageHeight = 256.0;
