+++ lib/page/marker/report_page.dart
@@ -2,17 +2,16 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:get_it_mixin/get_it_mixin.dart';
-import 'package:insigno_frontend/marker/add_images_widget.dart';
+import 'package:insigno_frontend/networking/authentication.dart';
 import 'package:insigno_frontend/networking/backend.dart';
 import 'package:insigno_frontend/networking/data/map_marker.dart';
 import 'package:insigno_frontend/networking/data/marker_type.dart';
+import 'package:insigno_frontend/page/map/location_provider.dart';
+import 'package:insigno_frontend/page/marker/add_images_widget.dart';
+import 'package:insigno_frontend/page/user/auth_user_provider.dart';
+import 'package:insigno_frontend/util/error_messages.dart';
 import 'package:insigno_frontend/util/error_text.dart';
-
-import '../map/location_provider.dart';
-import '../networking/authentication.dart';
-import '../user/auth_user_provider.dart';
-import '../util/error_messages.dart';
-import '../util/pair.dart';
+import 'package:insigno_frontend/util/pair.dart';
 
 class ReportPage extends StatefulWidget with GetItStatefulWidgetMixin {
   static const routeName = "/reportPage";
