+++ lib/page/marker/resolve_page.dart
@@ -2,18 +2,17 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:get_it_mixin/get_it_mixin.dart';
-import 'package:insigno_frontend/user/auth_user_provider.dart';
+import 'package:insigno_frontend/networking/authentication.dart';
+import 'package:insigno_frontend/networking/backend.dart';
+import 'package:insigno_frontend/networking/data/map_marker.dart';
+import 'package:insigno_frontend/page/map/location_provider.dart';
+import 'package:insigno_frontend/page/marker/add_images_widget.dart';
+import 'package:insigno_frontend/page/user/auth_user_provider.dart';
 import 'package:insigno_frontend/util/error_messages.dart';
 import 'package:insigno_frontend/util/error_text.dart';
 import 'package:insigno_frontend/util/nullable.dart';
+import 'package:insigno_frontend/util/pair.dart';
 import 'package:latlong2/latlong.dart';
-
-import '../map/location_provider.dart';
-import '../networking/authentication.dart';
-import '../networking/backend.dart';
-import '../networking/data/map_marker.dart';
-import '../util/pair.dart';
-import 'add_images_widget.dart';
 
 class ResolvePage extends StatefulWidget with GetItStatefulWidgetMixin {
   static const routeName = '/resolvePage';
