+++ lib/page/user/auth_user_provider.dart
@@ -2,8 +2,7 @@ import 'dart:async';
 
 import 'package:injectable/injectable.dart';
 import 'package:insigno_frontend/networking/backend.dart';
-
-import '../networking/data/authenticated_user.dart';
+import 'package:insigno_frontend/networking/data/authenticated_user.dart';
 
 @lazySingleton
 class AuthUserProvider {
