+++ lib/page/user/change_password_widget.dart
@@ -3,11 +3,10 @@ import 'package:flutter/services.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:get_it_mixin/get_it_mixin.dart';
 import 'package:insigno_frontend/networking/authentication.dart';
-import 'package:insigno_frontend/user/auth_user_provider.dart';
-import 'package:insigno_frontend/user/validators.dart';
+import 'package:insigno_frontend/networking/error.dart';
+import 'package:insigno_frontend/page/user/auth_user_provider.dart';
+import 'package:insigno_frontend/page/user/validators.dart';
 import 'package:insigno_frontend/util/error_text.dart';
-
-import '../networking/error.dart';
 
 class ChangePasswordWidget extends StatefulWidget with GetItStatefulWidgetMixin {
   final Function(bool changeRequestSent) finishCallback;
