+++ test/pill_page_test.dart
@@ -6,8 +6,8 @@
 // tree, read text, and verify that the values of widget properties are correct.
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:insigno_frontend/home/pill_page.dart';
 import 'package:insigno_frontend/networking/data/pill.dart';
+import 'package:insigno_frontend/page/pill_page.dart';
 
 void main() {
   testWidgets("Pill page test", (WidgetTester tester) async {
